+++ Gruntfile.js
@@ -199,6 +199,15 @@ module.exports = function (grunt) {
           ]
         }
       }
+    },
+    connect: {
+      server: {
+        options: {
+          port: 8080,
+          base: '.',
+          keepalive: true
+        }
+      }
     }
   });
 
@@ -213,4 +222,5 @@ module.exports = function (grunt) {
   grunt.loadNpmTasks('grunt-contrib-concat');
   grunt.loadNpmTasks('grunt-contrib-watch');
   grunt.loadNpmTasks('grunt-contrib-jasmine');
+  grunt.loadNpmTasks('grunt-contrib-connect');
 };