+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+Bugfixes:
+- autoColumnSize did not take `<table>` class into account when measuring the column width
+
+Performance:
+- use `cloneNode` in cell built-in cell renderers for better performance
+
+Other:
+- add `grunt-contrib-connect` to `Gruntfile.js` that allows to run an ad-hoc http://localhost:8080/ server with command `grunt connect`
+- bind `cellProperties` as `this` in `cells` callback
+
 ## [0.9.4](https://github.com/warpech/jquery-handsontable/tree/v0.9.4) (Jun 7, 2013)
 
 Bugfixes: